requirements.txt
fastapi
uvicorn[standard]
pydantic
python-docx
pymupdf
language_tool_python
langchain
langchain-community
langchain-huggingface
streamlit
spacy
groq
python-multipart